视频录制界面，引入第三方库

+++ LLama/LLama/UI/UserProfile/Controller/LLAUserAccountDetailViewController.m
@@ -362,7 +362,7 @@ static const NSInteger detailInfoSection = 2;
         
         withdrawCash = newInterger;
         //[dataTableView reloadRowsAtIndexPaths:@[[NSIndexPath indexPathForRow:0 inSection:withdrawCashSectionIndex]] withRowAnimation:UITableViewRowAnimationNone];
-        textField.text = [NSString stringWithFormat:@"%ld",newInterger];
+        textField.text = [NSString stringWithFormat:@"%ld",(long)newInterger];
     }
 }
